Item controller test working

+++ src/main/java/com/decagon/OakLandv1be/services/WalletService.java
@@ -4,10 +4,12 @@ import com.decagon.OakLandv1be.dto.FundWalletRequest;
 import com.decagon.OakLandv1be.dto.FundWalletResponseDto;
 import com.decagon.OakLandv1be.utils.ApiResponse;
 
+import java.math.BigDecimal;
+
 public interface WalletService {
 
     FundWalletResponseDto fundWallet(FundWalletRequest request);
 
-    Double getWalletBalance();
+    BigDecimal getWalletBalance();
 
 }
